src/alpha-vantage/requirements.txt
requests
python-dotenv
mysql-connector-python
